+++ CHANGELOG.md
@@ -62,6 +62,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Vulnerability: Debug Mode Enabled
 - Vulnerability: CVE-2022-26134
 - Vulnerability: Missing API Security Headers
+- Vulnerability: Path Traversal In API Route
 
 - FP: Information via "X-Powered-By" HTTP Response Header Field(s) Leaked By Server
 - FP: Dangerous JS Functions